# openapi.yaml
openapi: 3.0.0
info:
  title: 'Cube-3 API'
  contact:
    email: leo.paillard@gmail.com
  version: 1.0.0
servers:
  -
    url: 'http://cyber-dodo.fr:8080/api/v1'
    description: 'Cube-3 API'
paths:
...
  /customer:
    post:
      tags:
        - Customer
      summary: 'Add a new customer'
      operationId: 448b50e6bad3d11d9d527fe33366a61e
      requestBody:
        description: 'Customer data'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
      responses:
        '201':
          description: 'Customer created'
        '400':
          description: 'Invalid request data'
        '404':
          description: 'Company, User or CustomerStatus not found'
        '409':
          description: 'Customer already exists'
        '500':
          description: 'Internal server error'
  /customer/all:
    get:
      tags:
        - Customer
      summary: 'Get all customers'
      operationId: fcfb5653b2927c3fe119b62807cf78eb
      responses:
        '200':
          description: 'Customers found'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerResponse'
        '500':
          description: 'Internal server error'
  '/customer/{id}':
    get:
      tags:
        - Customer
      summary: 'Get customer by id'
      operationId: 0da629f9258ab86ebdb7b908bd3a9a7b
      parameters:
        -
          name: id
          in: path
          description: 'ID of the customer to retrieve'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Customer found'
        '404':
          description: 'Customer not found'
        '500':
          description: 'Internal server error'
    put:
      tags:
        - Customer
      summary: 'Update a customer'
      operationId: debaf18b1af5d04ee9fe39a6a4bc2676
      parameters:
        -
          name: id
          in: path
          description: 'ID of the customer to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: 'Customer data to update'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
      responses:
        '200':
          description: 'Customer updated successfully'
        '400':
          description: 'Invalid request data'
        '404':
          description: 'Customer not found'
        '409':
          description: 'Customer already exists'
        '500':
          description: 'Internal server error'
    delete:
      tags:
        - Customer
      summary: 'Delete a customer'
      operationId: 21d795818a35e5bb74096d73661c1b34
      parameters:
        -
          name: id
          in: path
          description: 'ID of the customer to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Customer deleted successfully'
        '404':
          description: 'Customer not found'
        '500':
          description: 'Internal server error'
  '/customer/company/{id}':
    get:
      tags:
        - Customer
      summary: 'Get customers by company id'
      operationId: 0729cb6591bfe79b21488a8d746ad726
      parameters:
        -
          name: id
          in: path
          description: 'ID of the company to retrieve customers from'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Customers found'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerResponse'
        '404':
          description: 'Customers not found'
        '500':
          description: 'Internal server error'
...
components:
  schemas:
...
    CompanyResponse:
      properties:
        id:
          type: integer
          example: '1'
        name:
          type: string
          example: 'Cube 3'
        address:
          type: string
          example: '1, rue de la paix'
        city:
          type: string
          example: Paris
        country:
          type: string
          example: France
        zipCode:
          type: string
          example: '75000'
        phone:
          type: string
          example: '0123456789'
        slogan:
          type: string
          example: 'The best company ever'
        logoPath:
          type: string
          example: cube3.png
        license:
          $ref: '#/components/schemas/LicenseResponse'
        language:
          type: string
          example: fr
        licenseExpirationDate:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
        createdAt:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
        updatedAt:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
      type: object
...
    CustomerRequest:
      required:
        - firstName
        - lastName
        - email
        - address
        - city
        - country
        - zipCode
        - phone
        - company
        - user
        - status
      properties:
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          format: email
          example: john-doe@gmail.com
        address:
          type: string
          example: '1, rue de la Paix'
        city:
          type: string
          example: Paris
        country:
          type: string
          example: France
        zipCode:
          type: string
          example: '75000'
        phone:
          type: string
          example: '0123456789'
        company:
          type: integer
          example: '1'
        user:
          type: integer
          example: '1'
        status:
          type: integer
          example: '1'
      type: object
    CustomerResponse:
      properties:
        id:
          type: integer
          example: '1'
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          format: email
          example: john-doe@gmail.com
        address:
          type: string
          example: '1, rue de la Paix'
        city:
          type: string
          example: Paris
        country:
          type: string
          example: France
        zipCode:
          type: string
          example: '75000'
        phone:
          type: string
          example: '0123456789'
        company:
          $ref: '#/components/schemas/CompanyResponse'
        user:
          $ref: '#/components/schemas/UserResponse'
        status:
          $ref: '#/components/schemas/CustomerStatusResponse'
        createdAt:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
        updatedAt:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
      type: object
...
    CustomerStatusResponse:
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'CustomerStatus 1'
        description:
          type: string
          example: 'This is the first customerStatus'
        createdAt:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
        updatedAt:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
      type: object
...
    LicenseResponse:
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: basic
        description:
          type: string
          example: 'basic license for 1 user'
        price:
          type: integer
          example: 100
        maxUsers:
          type: integer
          example: 1
        validityPeriod:
          type: integer
          example: 1
        createdAt:
          type: string
          format: date-time
          example: '2021-09-30 12:00:00'
        updatedAt:
          type: string
          format: date-time
          example: '2021-09-30 12:00:00'
      type: object
...
    RoleResponse:
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Admin
        description:
          type: string
          example: Administrator
        createdAt:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
        updatedAt:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
      type: object
...
    UserResponse:
      properties:
        id:
          type: integer
          example: '1'
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          example: john.doe@gmail.com
        job:
          type: string
          example: Developer
        phone:
          type: string
          example: '0123456789'
        role:
          $ref: '#/components/schemas/RoleResponse'
        company:
          type: integer
          example: '1'
        createdAt:
          type: string
          format: date-time
          example: '2021-03-01 00:00:00'
        updatedAt:
          type: string
          format: date-time
          example: '2021-03-01 00:00:00'
      type: object
...
  securitySchemes:
    bearer:
      type: http
      bearerFormat: JWT